# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Traffic Flow Analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  /** @nullable */
  details?: string | null;
}

export interface DatasetUploadInput {
  file: Blob;
}

export type DatasetProfileDataTypes = {[key: string]: string};

export type DatasetProfileClassDistribution = {[key: string]: number};

export type DatasetProfilePreviewItem = { [key: string]: unknown };

export interface DatasetProfile {
  datasetId: string;
  name: string;
  isDemo: boolean;
  rows: number;
  columns: number;
  columnNames: string[];
  dataTypes: DatasetProfileDataTypes;
  missingValues: number;
  duplicateRows: number;
  numericalFeatures: string[];
  categoricalFeatures: string[];
  /** @nullable */
  possibleTarget?: string | null;
  targetCandidates: string[];
  trafficColumns: string[];
  classDistribution: DatasetProfileClassDistribution;
  preview: DatasetProfilePreviewItem[];
}

export interface TrainModelsInput {
  datasetId: string;
  targetColumn: string;
  /**
     * @minimum 0.1
     * @maximum 0.4
     */
  testSize?: number;
}

export interface ModelMetric {
  name: string;
  accuracy: number;
  precision: number;
  recall: number;
  f1: number;
  trainingTimeMs: number;
}

export interface ConfusionMatrix {
  labels: string[];
  values: number[][];
}

export interface FeatureImportance {
  feature: string;
  importance: number;
  direction: string;
}

export type TrainingResultsClassDistribution = {[key: string]: number};

export interface TrainingResults {
  datasetId: string;
  targetColumn: string;
  bestModel: string;
  trainedAt: string;
  metrics: ModelMetric[];
  confusionMatrix: ConfusionMatrix;
  featureImportance: FeatureImportance[];
  classDistribution: TrainingResultsClassDistribution;
}

export type PredictionInputFeatures = { [key: string]: unknown };

export interface PredictionInput {
  datasetId: string;
  features: PredictionInputFeatures;
}

export type PredictionResultInputSummary = {[key: string]: string | number | boolean | null};

export interface PredictionResult {
  datasetId: string;
  condition: string;
  confidence: number;
  model: string;
  factors: FeatureImportance[];
  inputSummary: PredictionResultInputSummary;
}

export interface AnalysisSummary {
  profile: DatasetProfile;
  training: TrainingResults | null;
  latestPrediction: PredictionResult | null;
}

export interface RagQueryInput {
  /** @minLength 1 */
  question: string;
  /** @nullable */
  datasetId?: string | null;
  prediction?: PredictionResult | null;
}

export interface RagSource {
  id: string;
  title: string;
  topic: string;
  excerpt: string;
}

export interface RagQueryResult {
  answer: string;
  retrievedFacts: string[];
  modelContext: string[];
  sources: RagSource[];
}

export type DatasetIdParameter = string;

export type GetModelResultsParams = {
datasetId: DatasetIdParameter;
};

export type GetAnalysisSummaryParams = {
datasetId: DatasetIdParameter;
};