# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Traffic Flow Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { DatasetProfile } from './datasetProfile';
import type { PredictionResult } from './predictionResult';
import type { TrainingResults } from './trainingResults';

export interface AnalysisSummary {
  profile: DatasetProfile;
  training: TrainingResults | null;
  latestPrediction: PredictionResult | null;
}