# lib/api-zod/src/generated/types/confusionMatrix.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Traffic Flow Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface ConfusionMatrix {
  labels: string[];
  values: number[][];
}