# lib/api-zod/src/generated/types/featureImportance.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Traffic Flow Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface FeatureImportance {
  feature: string;
  importance: number;
  direction: string;
}