# lib/api-zod/src/generated/types/predictionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Traffic Flow Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionInputFeatures } from './predictionInputFeatures';

export interface PredictionInput {
  datasetId: string;
  features: PredictionInputFeatures;
}