# lib/api-zod/src/generated/types/ragQueryInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Traffic Flow Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionResult } from './predictionResult';

export interface RagQueryInput {
  /** @minLength 1 */
  question: string;
  /** @nullable */
  datasetId?: string | null;
  prediction?: PredictionResult | null;
}