# lib/api-zod/src/generated/types/ragQueryResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Traffic Flow Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { RagSource } from './ragSource';

export interface RagQueryResult {
  answer: string;
  retrievedFacts: string[];
  modelContext: string[];
  sources: RagSource[];
}