# lib/api-zod/src/generated/types/ragSource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Traffic Flow Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface RagSource {
  id: string;
  title: string;
  topic: string;
  excerpt: string;
}