# lib/api-zod/src/generated/types/trainModelsInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Traffic Flow Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface TrainModelsInput {
  datasetId: string;
  targetColumn: string;
  /**
     * @minimum 0.1
     * @maximum 0.4
     */
  testSize?: number;
}