# lib/api-zod/src/generated/types/trainingResults.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Traffic Flow Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { ConfusionMatrix } from './confusionMatrix';
import type { FeatureImportance } from './featureImportance';
import type { ModelMetric } from './modelMetric';
import type { TrainingResultsClassDistribution } from './trainingResultsClassDistribution';

export interface TrainingResults {
  datasetId: string;
  targetColumn: string;
  bestModel: string;
  trainedAt: string;
  metrics: ModelMetric[];
  confusionMatrix: ConfusionMatrix;
  featureImportance: FeatureImportance[];
  classDistribution: TrainingResultsClassDistribution;
}